requirements.txt
# Flask and extensions
flask==2.0.3
flask-cors==3.0.10
flask-jwt-extended==4.3.1
flask-session==0.4.0

# MongoDB
pymongo[srv]==4.0.1

# Environment variables
python-dotenv==0.19.1

# Security
bcrypt==4.0.1
pyjwt==2.6.0
pydantic==2.5.0
email-validator==2.1.0

# Utilities
requests==2.26.0
python-dateutil==2.8.2
dnspython==2.1.0

# PDF generation
reportlab==4.0.7

# Production server
gunicorn==20.1.0
werkzeug==2.0.3

# Groq AI API
groq==0.4.1

# For deployment
vercel==0.2.1

# Testing and development
pytest==6.2.5
flake8==4.0.1